+++ frontend/widgets/views/subscribe-form.php
@@ -1,7 +1,6 @@
 <?php
 use yii\helpers\Html;
-use yii\bootstrap\BootstrapAsset;
-use yii\bootstrap\ActiveForm;
+use yii\widgets\ActiveForm;
 
 /**
  * View file for Subscribe widget
@@ -15,19 +14,17 @@ use yii\bootstrap\ActiveForm;
  * @license BSD 3-Clause License
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  */
-
-BootstrapAsset::register($this);
 ?>
 
 <?php
 $form = ActiveForm::begin([
     'action' => $actionRoute,
-    'options' => ['class' => 'form-inline'],
     'enableClientValidation' => false
 ])
 ?>
     <?= $form->field($model, 'email')
-            ->textInput()
-            ->label(false) ?>
+            ->textInput() ?>
+    <?= $form->field($model, 'phone')
+            ->textInput() ?>
     <?= Html::submitButton('Subscribe', ['class' => 'btn btn-success']) ?>
 <?php $form->end() ?>